16 sprint - Search block refactored
