--- app/src/main/java/com/practicum/playlistmaker/data/dto/TracksResponse.kt
@@ -1,5 +1,0 @@
-package com.practicum.playlistmaker.data.dto
-
-import com.practicum.playlistmaker.domain.entities.Track
-
-class TracksResponse (val results: ArrayList<Track>)
