+++ app/src/main/java/com/practicum/playlistmaker/data/search/NetworkClient.kt
@@ -1,0 +1,7 @@
+package com.practicum.playlistmaker.data.search
+
+import com.practicum.playlistmaker.data.search.models.NetworkResponse
+
+interface NetworkClient {
+    fun doRequest(requestModel: Any): NetworkResponse
+}
