+++ app/src/main/java/com/practicum/playlistmaker/data/search/models/NetworkResponse.kt
@@ -1,0 +1,5 @@
+package com.practicum.playlistmaker.data.search.models
+
+open class NetworkResponse() {
+    var resultCode = 0
+}
