+++ app/src/main/java/com/practicum/playlistmaker/data/search/models/TracksSearchRequest.kt
@@ -1,0 +1,3 @@
+package com.practicum.playlistmaker.data.search.models
+
+data class TracksSearchRequest(val expression: String)
