+++ app/src/main/java/com/practicum/playlistmaker/data/search/models/TracksSearchResponse.kt
@@ -1,0 +1,3 @@
+package com.practicum.playlistmaker.data.search.models
+
+data class TracksSearchResponse (val results: ArrayList<TrackDto>): NetworkResponse()
