+++ app/src/main/java/com/practicum/playlistmaker/domain/player/PlayerInteractor.kt
@@ -1,0 +1,18 @@
+package com.practicum.playlistmaker.domain.player
+
+import com.practicum.playlistmaker.domain.player.model.PlayerState
+import com.practicum.playlistmaker.domain.search.models.Track
+
+interface PlayerInteractor {
+    fun prepare(listener: OnStateChangeListener)
+    fun start()
+    fun pause()
+    fun release()
+    fun getPlayerCurrentTimerPosition(): String
+    fun getPlayerState(): PlayerState
+    fun setPlayerDataSource(track: Track)
+
+    interface OnStateChangeListener {
+        fun onChange(state: PlayerState)
+    }
+}
