+++ app/src/main/java/com/practicum/playlistmaker/domain/player/model/PlayerState.kt
@@ -1,0 +1,8 @@
+package com.practicum.playlistmaker.domain.player.model
+
+enum class PlayerState {
+    DEFAULT,
+    PREPARED,
+    PLAYING,
+    PAUSED,
+}
