--- app/src/main/java/com/practicum/playlistmaker/presentation/ui/mediateka/MediatekaActivity.kt
@@ -1,14 +1,0 @@
-package com.practicum.playlistmaker.presentation.ui.mediateka
-
-import androidx.appcompat.app.AppCompatActivity
-import android.os.Bundle
-import com.practicum.playlistmaker.databinding.ActivityMediatekaBinding
-
-class MediatekaActivity : AppCompatActivity() {
-    private lateinit var binding: ActivityMediatekaBinding
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        binding = ActivityMediatekaBinding.inflate(layoutInflater)
-        setContentView(binding.root)
-    }
-}
