+++ app/src/main/res/layout/activity_mediateka.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/white"
-    tools:context=".presentation.ui.mediateka.MediatekaActivity">
+    tools:context=".ui.mediateka.MediatekaActivity">
 
     <TextView
         android:id="@+id/message"
