+++ app/src/main/res/layout/activity_player.xml
@@ -9,7 +9,7 @@
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
-        tools:context=".presentation.ui.player.PlayerActivity">
+        tools:context=".ui.player.activity.PlayerActivity">
 
         <androidx.appcompat.widget.Toolbar
             android:id="@+id/tbBackFromPlayerButton"
