+++ app/src/main/res/layout/activity_search.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     style="@style/SearchScreenLayout"
     android:orientation="vertical"
-    tools:context=".presentation.ui.search.SearchActivity">
+    tools:context=".ui.search.activity.SearchActivity">
 
     <androidx.appcompat.widget.Toolbar
         android:id="@+id/searchOutButton"
