+++ app/src/main/res/layout/activity_settings.xml
@@ -2,7 +2,7 @@
 <LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
-    tools:context=".presentation.ui.settings.SettingsActivity"
+    tools:context=".ui.settings.SettingsActivity"
     style="@style/SettingsScreenLayout">
 
     <Button
